ext2spice/ext2spice.c: Multiplication result converted to larger type
Fix code scanning alert no. 43: Multiplication result converted to larger type (#59)

+++ ext2spice/ext2spice.c
@@ -2011,8 +2011,8 @@ spcWriteParams(dev, hierName, scale, l, w, sdM)
 		    if (esScale < 0)
 			fprintf(esSpiceF, "%g", parmval * scale * scale);
 		    else if (plist->parm_scale != 1.0)
-			fprintf(esSpiceF, "%g", parmval * scale * scale
-				* esScale * esScale * plist->parm_scale
+			fprintf(esSpiceF, "%g", (double)parmval * (double)scale * (double)scale
+				* (double)esScale * (double)esScale * plist->parm_scale
 				* 1E-12);
 		    else
 			esSIvalue(esSpiceF, 1.0E-12 * (parmval + plist->parm_offset)
